requirements.txt
fastapi~=0.115.6
uvicorn
sqlalchemy~=2.0.36
pyjwt
requests~=2.32.3
pydantic~=2.10.4
google-auth
google-auth-httplib2
google-api-python-client
python-dateutil
pandas